chore: revise notes, comments, and mocks
